# docs/superpowers/specs/idea-factory-openapi.yaml
openapi: 3.1.0
info:
  title: Idea Factory API
  version: 1.0.0-target
  description: |
    Target-state API contract for workspace governance and MainAgent-driven graph runtime.
servers:
  - url: /v1
...
paths:
  /workspaces:
    post:
      tags: [Workspaces]
      summary: Create workspace
      operationId: createWorkspace
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: Get workspace
      operationId: getWorkspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '404': { $ref: '#/components/responses/NotFound' }
...
  /workspaces/{workspaceId}/runtime:
    get:
      tags: [Runs]
      summary: Get runtime state
      operationId: getRuntimeState
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: run_id
          in: query
          schema: { type: string }
        - name: latest_runs
          in: query
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Runtime state snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeStateResponse'
        '404': { $ref: '#/components/responses/NotFound' }
...
  /workspaces/{workspaceId}/interventions/{interventionId}/events:
    get:
      tags: [Interventions]
      summary: List intervention lifecycle events
      operationId: listInterventionEvents
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/InterventionId'
        - name: cursor
          in: query
          schema: { type: string }
          description: Cursor from previous page in format `RFC3339|event_id`.
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/InterventionStatus'
      responses:
        '200':
          description: Intervention lifecycle events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionEventsResponse'
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string }
...
    InterventionId:
      name: interventionId
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    WorkspaceStatus:
      type: string
      enum: [draft, active, paused, archived]
...
    InterventionStatus:
      type: string
      enum: [received, absorbed, reflected, rejected]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_argument, not_found, conflict, too_many_requests, internal]
            message:
              type: string
            details:
              type: array
              items:
                type: string
    CreateWorkspaceRequest:
      type: object
      required: [topic, goal]
      properties:
        topic: { type: string, minLength: 1 }
        goal: { type: string, minLength: 1 }
        constraints:
          type: array
          items: { type: string }
        budget:
          type: object
          properties:
            max_runs: { type: integer, minimum: 1 }
            timebox_minutes: { type: integer, minimum: 1 }
    WorkspaceResponse:
      type: object
      required: [workspace]
      properties:
        workspace:
          type: object
          required: [id, topic, goal, status, created_at, updated_at]
          properties:
            id: { type: string }
            topic: { type: string }
            goal: { type: string }
            constraints:
              type: array
              items: { type: string }
            status:
              $ref: '#/components/schemas/WorkspaceStatus'
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    RuntimeStateResponse:
      type: object
      required: [runtime]
      properties:
        runtime:
          $ref: '#/components/schemas/RuntimeState'
    RuntimeState:
      type: object
      required: [runs, agent_tasks, results, balance]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeRun'
        agent_tasks:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeAgentTask'
        results:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeTaskResult'
        mutations:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeMutation'
        balance:
          $ref: '#/components/schemas/BalanceState'
        latest_replan_reason:
          type: string
          nullable: true
    RuntimeRun:
      type: object
      required: [id, workspace_id, source, status, started_at]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        source:
          type: string
          enum: [manual, auto, resume, workspace_create, workspace_load, workspace_read]
        status:
          type: string
          enum: [pending, running, completed, failed]
        started_at: { type: integer }
        ended_at: { type: integer, nullable: true }
    RuntimeAgentTask:
      type: object
      required: [id, workspace_id, run_id, sub_agent, goal, status, updated_at]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        run_id: { type: string }
        sub_agent: { type: string }
        goal: { type: string }
        status:
          type: string
          enum: [todo, doing, done, failed, skipped]
        updated_at: { type: integer }
    RuntimeTaskResult:
      type: object
      required: [task_id, summary, is_success, updated_at]
      properties:
        task_id: { type: string }
        summary: { type: string }
        is_success: { type: boolean }
        updated_at: { type: integer }
    RuntimeMutation:
      type: object
      required: [id, workspace_id, kind, created_at]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        kind:
          type: string
          enum:
            - run_created
            - run_completed
            - run_failed
            - node_added
            - edge_added
            - intervention_absorbed
            - balance_updated
            - active_opportunity_set
            - favorites_updated
            - strategy_updated
        source: { type: string, nullable: true }
        created_at: { type: integer }
    BalanceState:
      type: object
      required: [workspace_id, run_id, divergence, research, aggression, reason, updated_at]
      properties:
        workspace_id: { type: string }
        run_id: { type: string }
        divergence: { type: number }
        research: { type: number }
        aggression: { type: number }
        reason: { type: string }
        updated_at: { type: integer }
...
    InterventionEvent:
      type: object
      required: [id, intervention_id, workspace_id, status, created_at]
      properties:
        id: { type: string }
        intervention_id: { type: string }
        workspace_id: { type: string }
        status: { $ref: '#/components/schemas/InterventionStatus' }
        created_at: { type: string, format: date-time }
    InterventionEventsResponse:
      type: object
      required: [workspace_id, intervention_id, events, has_more]
      properties:
        workspace_id: { type: string }
        intervention_id: { type: string }
        events:
          type: array
          items:
            $ref: '#/components/schemas/InterventionEvent'
        next_cursor: { type: string }
        has_more: { type: boolean }
...